update to 64-bit and fix all warnings

+++ src/AtomicParsley.h
@@ -205,7 +205,7 @@ void APar_Unified_atom_Put(AtomicInfo* target_atom, const char* unicode_data,
   uint8_t text_tag_style, uint64_t ancillary_data, uint8_t anc_bit_width);
 
 void APar_atom_Binary_Put(AtomicInfo* target_atom, const char* binary_data,
-  uint32_t bytecount, uint64_t atomic_data_offset);
+  uint64_t bytecount, uint64_t atomic_data_offset);
 
 /* iTunes-style metadata */
 void APar_MetaData_atomArtwork_Set(const char* artworkPath,
@@ -233,7 +233,7 @@ uint16_t APar_TestVideoDescription(AtomicInfo* video_desc_atom,
 void APar_Generate_iPod_uuid(char* atom_path);
 
 /* 3GP-style metadata */
-uint32_t APar_3GP_Keyword_atom_Format(char* keywords_globbed,
+uint64_t APar_3GP_Keyword_atom_Format(char* keywords_globbed,
   uint8_t keyword_count, bool set_UTF16_text, char* &formed_keyword_struct);
 
 AtomicInfo* APar_UserData_atom_Init(const char* userdata_atom_name,
@@ -256,11 +256,9 @@ void APar_DetermineAtomLengths();
 void APar_WriteFile(const char* ISObasemediafile, const char* outfile,
   bool rewrite_original);
 
-void APar_zlib_inflate(char* in_buffer, uint32_t in_buf_len,
-  char* out_buffer, uint32_t out_buf_len);
+void APar_zlib_inflate(char* in_buffer, uint32_t in_buf_len, char* out_buffer, uint32_t out_buf_len);
 
-uint32_t APar_zlib_deflate(char* in_buffer, uint32_t in_buf_len,
-  char* out_buffer, uint32_t out_buf_len);
+uint64_t APar_zlib_deflate(char* in_buffer, uint32_t in_buf_len, char* out_buffer, uint32_t out_buf_len);
 
 
 void APar_print_uuid(ap_uuid_t* uuid, bool new_line = true);
@@ -320,27 +318,21 @@ extern int sha1_stream (FILE *stream, void *resblock);
    digest.  */
 extern void *sha1_buffer (const char *buffer, size_t len, void *resblock);
 
-int isolat1ToUTF8(unsigned char* out, int outlen,
-  const unsigned char* in, int inlen);
+size_t isolat1ToUTF8(unsigned char* out, size_t outlen,  const unsigned char* in, size_t inlen);
 
-int UTF8Toisolat1(unsigned char* out, int outlen,
-  const unsigned char* in, int inlen);
+size_t UTF8Toisolat1(unsigned char* out, size_t outlen, const unsigned char* in, size_t inlen);
 
-int UTF16BEToUTF8(unsigned char* out, int outlen,
-  const unsigned char* inb, int inlenb);
+size_t UTF16BEToUTF8(unsigned char* out, size_t outlen, const unsigned char* inb, size_t inlenb);
 
-int UTF8ToUTF16BE(unsigned char* outb, int outlen,
-  const unsigned char* in, int inlen);
+size_t UTF8ToUTF16BE(unsigned char* outb, size_t outlen, const unsigned char* in, size_t inlen);
 
-int UTF16LEToUTF8(unsigned char* out, int outlen,
-  const unsigned char* inb, int inlenb);
+size_t UTF16LEToUTF8(unsigned char* out, size_t outlen, const unsigned char* inb, size_t inlenb);
 
-int UTF8ToUTF16LE(unsigned char* outb, int outlen,
-  const unsigned char* in, int inlen);
+size_t UTF8ToUTF16LE(unsigned char* outb, size_t outlen, const unsigned char* in, size_t inlen);
 
 int isUTF8(const char* in_string);
 
-unsigned int utf8_length(const char *in_string, unsigned int char_limit);
+size_t utf8_length(const char *in_string, size_t char_limit);
 
 int strip_bogusUTF16toRawUTF8(unsigned char* out, int inlen,
   wchar_t* in, int outlen);
